requirements.txt
mysql-connector-python
streamlit
pandas
plotly